requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dropbox==12.0.2
fastapi==0.112.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
notion-client==2.2.1
ply==3.11
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==0.18.0
pytz==2021.3
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
stone==3.3.1
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
